pyproject.toml
[project]
name = "upload-artifacts"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "gitpython==3.1.46",
    "minio==7.2.20",
]
